fix!: popups now render dynamically :)
ftxui::Modal() doesn't redraw, it takes it's first child component and holds onto it with diamond hands. unlike renderer(), can't pass a pointer to it and expect it to draw the change itself. since Modal() is ComponentDecorator and not a Component itself. Had to rig a collapsible Container::Horizontal({}) to hold popupContent dynamically. FTXUI has been pretty intuitive as a functional library but i was stuck on this for approximately 4 hours straight. just staring at nvim, documentation and github issues. ARTHUR SONGONZI UPDATE YOUR DOCUMENTATION YOU LITTLE FRENCHMAN..